Reuse serviceValidate for proxyValidate
